Updated quite a few things.
